Feature/auth flow (#2)
* feat: use UserDTO instead of User

* feat: register user

* feat: add auth endpoints

* feat: add jwtguard to users controller

* refactor: forbidden exception on jwt strategy

* refactor: remove app service

+++ apps/api/src/modules/users/use-cases/delete-user/index.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@nestjs/common';
-import { DeleteUserParamsDTO } from '@repo/api';
+import { ForbiddenException, Injectable } from '@nestjs/common';
+import { DeleteUserParamsDTO, UserDTO } from '@repo/api';
 import UserRepository from '~/services/database/typeorm/repositories/user.repository';
 import { UseCase } from '~/shared/core/use-case';
 
-type DeleteUserParams = DeleteUserParamsDTO;
+type DeleteUserParams = DeleteUserParamsDTO & { currentUser: UserDTO };
 type DeleteUserResult = boolean;
 
 @Injectable()
@@ -12,8 +12,17 @@ export default class DeleteUser
 {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async execute({ id }: DeleteUserParams): Promise<DeleteUserResult> {
-    // TODO: Check id is the same of the currentUser or currentUser is admin
+  async execute({
+    id,
+    currentUser,
+  }: DeleteUserParams): Promise<DeleteUserResult> {
+    const isSameUser = currentUser.id === id;
+    if (!isSameUser) {
+      throw new ForbiddenException(
+        'You can only edit your own account information',
+      );
+    }
+    // TODO: Or currentUser is admin
     return this.userRepository.delete(id);
   }
 }
